+++ packages/@sanity/cli/test/helpers/waitForAsync.ts
@@ -1,0 +1,3 @@
+export const waitForAsync = (ms = 0): Promise<void> => {
+  return new Promise((resolve) => setTimeout(resolve, ms))
+}
